% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plots.R
\name{plot_mixomics_perf}
\alias{plot_mixomics_perf}
\title{Plot PLS performance}
\usage{
plot_mixomics_perf(perf_pls, ncomp)
}
\arguments{
\item{ncomp}{number of X components}

\item{model}{a PLS or sPLS model}
}
\description{
A helper function for \code{mixomics_pls} and \code{mixomics_spls}.
}
